requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
openai>=1.3.0
pillow>=9.5.0
requests==2.31.0
python-dotenv==1.0.0
werkzeug==2.3.7
httpx==0.25.0